additional test fixes and cleanup
