+++ liquibase-integration-tests/src/test/groovy/liquibase/snapshot/JdbcDatabaseSnapshotH2IntegrationTest.groovy
@@ -1,30 +1,25 @@
 package liquibase.snapshot
 
-
 import liquibase.Scope
-import liquibase.command.util.CommandUtil
 import liquibase.database.DatabaseFactory
 import liquibase.database.jvm.JdbcConnection
-import liquibase.extension.testing.command.CommandTests
 import liquibase.extension.testing.testsystem.DatabaseTestSystem
 import liquibase.extension.testing.testsystem.TestSystemFactory
-import liquibase.resource.SearchPathResourceAccessor
+import liquibase.extension.testing.testsystem.spock.LiquibaseIntegrationTest
 import liquibase.statement.SqlStatement
 import liquibase.statement.core.RawSqlStatement
 import liquibase.structure.core.Column
 import liquibase.structure.core.Table
 import liquibase.structure.core.View
-import org.junit.Rule
+import spock.lang.Shared
 import spock.lang.Specification
 import spock.lang.Unroll
 
-class JdbcDatabaseSnapshotIntegrationTest extends Specification {
+@LiquibaseIntegrationTest
+class JdbcDatabaseSnapshotH2IntegrationTest extends Specification {
 
-    @Rule
+    @Shared
     public DatabaseTestSystem h2 = Scope.currentScope.getSingleton(TestSystemFactory).getTestSystem("h2")
-
-    @Rule
-    public DatabaseTestSystem mssql = Scope.currentScope.getSingleton(TestSystemFactory).getTestSystem("mssql")
 
     @Unroll
     def "getTables, getColumns and getViews works with underscores in schema names"() {
@@ -53,27 +48,5 @@ class JdbcDatabaseSnapshotIntegrationTest extends Specification {
                 new RawSqlStatement("drop schema \"test_schema\" if exists"),
                 new RawSqlStatement("drop schema \"test-schema\" if exists"),
         ] as SqlStatement[], null)
-    }
-
-    def "columnstore indexes are deserialized out of snapshot json files properly"() {
-        when:
-        mssql.executeSql('CREATE TABLE address(id1 int, id2 int, id3 int)')
-        mssql.executeSql('CREATE COLUMNSTORE INDEX IX_address ON address(id1, id2, id3)')
-        def snapshotFile = CommandTests.takeDatabaseSnapshot(mssql.getDatabaseFromFactory(), "json")
-        def outputFile = 'output.mssql.sql'
-        // The results of this operation are unimportant. All this serves to do is to cause Liquibase to deserialize
-        // the snapshot file.
-        Scope.child(Collections.singletonMap(Scope.Attr.resourceAccessor.name(), new SearchPathResourceAccessor(".,target/test-classes/")), new Scope.ScopedRunner() {
-            @Override
-            void run() throws Exception {
-                CommandUtil.runDiff(mssql.getDatabaseFromFactory(), 'offline:sqlserver?snapshot=' + snapshotFile.toString(), outputFile)
-            }
-        })
-
-        then:
-        noExceptionThrown()
-
-        cleanup:
-        CommandUtil.runDropAll(mssql)
     }
 }
